update to electron 12.0.0 ; made necessary changes to config files, imports, and ipc communication to support
