+++ package.json
@@ -28,8 +28,9 @@
     "clean:releases": "npx del releases",
     "clean": "npm run clean:dist && npm run clean:releases",
     "compile:renderer": "webpack --config webpack.renderer.js",
+    "compile:preload": "webpack --config webpack.preload.js",
     "compile:main": "webpack --config webpack.main.js",
-    "compile": "npm run clean:deps && npm run compile:renderer && npm run compile:main",
+    "compile": "npm run clean && npm run compile:renderer && npm run compile:preload && npm run compile:main",
     "dist": "npm run compile && electron-builder",
     "build:mac": "npm run compile && electron-builder --mac",
     "build:win": "npm run compile && electron-builder --win",
@@ -43,6 +44,7 @@
     "extends": null,
     "files": [
       "dist/main/**/*",
+      "dist/preload/**/*",
       "dist/renderer/**/*",
       "dist/renderer/index.html",
       "dist/renderer/renderer.js"
@@ -91,12 +93,12 @@
     "prosemirror-transform": "^1.2.7",
     "prosemirror-view": "^1.15.2",
     "solid-js": "^0.18.14",
-    "source-map-support": "^0.5.19",
     "yaml": "^1.10.0"
   },
   "devDependencies": {
+    "@babel/core": "^7.13.8",
     "@babel/plugin-proposal-optional-chaining": "^7.10.4",
-    "@babel/preset-env": "^7.10.4",
+    "@babel/preset-env": "^7.13.9",
     "@babel/preset-typescript": "^7.10.4",
     "@types/electron-store": "^3.2.0",
     "@types/katex": "^0.11.0",
@@ -114,17 +116,27 @@
     "@types/prosemirror-tables": "^0.9.1",
     "@types/prosemirror-transform": "^1.1.1",
     "@types/prosemirror-view": "^1.15.0",
+    "@types/source-map-support": "^0.5.3",
     "asar": "^3.0.3",
+    "babel-loader": "^8.2.2",
     "babel-preset-solid": "^0.18.14",
     "copy-webpack-plugin": "^6.0.3",
-    "electron": "^9.1.1",
+    "css-loader": "^5.1.1",
+    "del": "^6.0.0",
+    "del-cli": "^3.0.1",
+    "electron": "^12.0.0",
     "electron-builder": "^22.7.0",
-    "electron-webpack": "^2.8.2",
-    "electron-webpack-ts": "^4.0.1",
+    "file-loader": "^6.2.0",
+    "html-loader": "^1.3.2",
     "html-webpack-plugin": "^4.5.2",
     "less-loader": "^6.2.0",
+    "source-map-support": "^0.5.19",
+    "ts-loader": "^6.2.2",
     "tsconfig-paths-webpack-plugin": "^3.2.0",
     "typescript": "^4.2.2",
-    "webpack": "^4.44.0"
+    "url-loader": "^4.1.1",
+    "webpack": "^4.44.0",
+    "webpack-cli": "^4.5.0",
+    "webpack-merge": "^5.7.3"
   }
 }
