+++ src/main/plugins/outline-plugin.ts
@@ -5,7 +5,7 @@ import path from "path";
 import { Node as ProseNode, Mark } from "prosemirror-model";
 
 // project imports
-import { IWorkspaceDir, IFileMeta } from "@common/fileio";
+import { IWorkspaceDir, IFileMeta } from "@common/files";
 import { WorkspacePlugin } from "./plugin";
 import { IDoc } from "@common/doctypes/doctypes";
