+++ src/main/workspace/workspace.ts
@@ -3,7 +3,8 @@ import * as pathlib from "path";
 import fs from "fs";
 
 // project imports
-import { IFileMeta, IFileDesc, IDirectory, IDirEntryMeta, readFile } from "@common/fileio";
+import { IFileMeta, IFileDesc, IDirectory, IDirEntryMeta } from "@common/files";
+import { readFile } from "@common/fileio";
 import { IDisposable } from "@common/types";
 import { ChokidarEvents } from "@common/events";
 import { IDoc } from "@common/doctypes/doctypes";
