+++ src/renderer/editors/editor-markdown.tsx
@@ -1,6 +1,3 @@
-// electron imports
-import { clipboard } from "electron";
-
 // prosemirror imports
 import { EditorView as ProseEditorView, EditorView } from "prosemirror-view";
 import { Mark, Slice, Node as ProseNode } from "prosemirror-model";
@@ -10,7 +7,7 @@ import { history, undo, redo } from "prosemirror-history";
 import { gapCursor } from "prosemirror-gapcursor";
 
 // project imports
-import { IPossiblyUntitledFile } from "@common/fileio";
+import { IPossiblyUntitledFile } from "@common/files";
 import { Editor } from "./editor";
 
 // markdown
@@ -342,27 +339,28 @@ export class MarkdownEditor extends Editor<ProseEditorState> {
 			handlePaste: (view) => {
 
 				/** @todo (6/22/20) make this work with the ClipboardEvent? */
+				/** @todo (2021/03/05) this was temporarily removed while setting up `contextIsolation` */
 
 				// for some reason, event.clipboardData.getData("img/png") etc.
 				// do not return any data.  So we use the electron clipboard instead.
-				if(clipboard.availableFormats("clipboard").find(str => str.startsWith("image"))){
-					let dataUrl:string = clipboard.readImage("clipboard").toDataURL();
+				// if(clipboard.availableFormats("clipboard").find(str => str.startsWith("image"))){
+				// 	let dataUrl:string = clipboard.readImage("clipboard").toDataURL();
 					
-					let imgNode = this._imageExt.type.createAndFill({
-						src: dataUrl
-					});
+				// 	let imgNode = this._imageExt.type.createAndFill({
+				// 		src: dataUrl
+				// 	});
 					
-					if(imgNode){
-						let { $from } = view.state.selection;
-						let tr = view.state.tr.deleteSelection().insert(
-							$from.pos,
-							imgNode
-						)
-						view.dispatch(tr);
-						return true;
-					}
+				// 	if(imgNode){
+				// 		let { $from } = view.state.selection;
+				// 		let tr = view.state.tr.deleteSelection().insert(
+				// 			$from.pos,
+				// 			imgNode
+				// 		)
+				// 		view.dispatch(tr);
+				// 		return true;
+				// 	}
 
-				}
+				// }
 				
 				return false;
 			}
