+++ src/renderer/index.ts
@@ -1,8 +1,11 @@
 import Renderer from './render';
 
 // other imports
-import { Titlebar, Color } from "custom-electron-titlebar";
+//import { Titlebar, Color } from "custom-electron-titlebar";
 import path from "path";
+
+// source-map-support
+import 'source-map-support/register'
 
 // library css
 import "@lib/prosemirror-math/style/math.css";
@@ -31,11 +34,13 @@ onload = function(){
 
 	/* HOT MODULE REPLACEMENT */
 
-	if (module.hot) {
+	// TODO: (2021/03/04) re-enable hot module reloading, which was disabled when migrating from electron-webpack
+
+	/*if (module.hot) {
 
 		module.hot.accept('./render', () => {
 			require('./render').default();
 		});
 
-	}
+	}*/
 }
