+++ webpack.base.js
@@ -2,10 +2,9 @@
 // (https://github.com/notable/notable/blob/54646c1fb64fbcf3cc0857a2791cfb6a6ae48313/webpack.base.js)
 
 const TSConfigPathsPlugin = require("tsconfig-paths-webpack-plugin");
-const CopyWebpackPlugin = require('copy-webpack-plugin');
 const path = require("path");
 const webpack = require("webpack");
-const merge = require("webpack-merge");
+const { merge } = require("webpack-merge");
 
 const PACKAGE_JSON = require("./package.json");
 
@@ -29,58 +28,21 @@ function PluginSkeletonOptimization ( compiler ) { // Loading heavy resources af
   });
 }
 
-// TODO: (2021/3/21) this was blindly copied from electron-webpack, but I should make an effort to replace+simplify it later
-// https://github.com/electron-userland/electron-webpack/blob/8a9d2892ecedd8f0d8158f3175116f34efdf94ed/packages/electron-webpack/src/main.ts#L278
-function computeExternals(isRenderer /*bool*/) {
-	// electron-webpack code was written with consideration for other build targets,
-	// but for our purposes, if we're not the renderer we MUST be the main process
-	let isMain = !isRenderer;
-
-	// whitelisted modules
-	const manualWhiteList = PACKAGE_JSON.electronWebpack.whiteListedModules;
-	const whiteListedModules = new Set(manualWhiteList || [])
-
-	if (isRenderer) {
-		whiteListedModules.add("react")
-		whiteListedModules.add("react-dom")
-	}
-
-	const filter = (name) => !name.startsWith("@types/") && (whiteListedModules == null || !whiteListedModules.has(name))
-	const externals = Object.keys(PACKAGE_JSON.dependencies).filter(filter)
-	externals.push("electron")
-	externals.push("webpack")
-	// because electron-devtools-installer specified in the devDependencies, but required in the index.dev
-	externals.push("electron-devtools-installer")
-	if (isMain) {
-		externals.push("webpack/hot/log-apply-result")
-		externals.push("electron-webpack/out/electron-main-hmr/HmrClient")
-		externals.push("source-map-support/source-map-support.js")
-	}
-
-	// if (this.electronWebpackConfiguration.externals != null) {
-	// 	return externals.concat(this.electronWebpackConfiguration.externals)
-	// }
-
-	return externals;
-}
-
 ////////////////////////////////////////////////////////////////////////////////
 
 function base(options){
 	// modify default config with a config function
 	// (https://webpack.electron.build/modifying-webpack-configurations#using-a-config-function)
 	let IS_PRODUCTION = (process.env.NODE_ENV !== "development");
-	let isRenderer = (options.target == "electron-renderer");
-	let DIRNAME = path.resolve(__dirname);
+	//let isRenderer = (options.target == "electron-renderer");
 	return (env) => {
 		// merge incoming config with a few extra options
-		let result = merge(options, {
+		let result = merge({
 			mode: IS_PRODUCTION ? "production" : "development",
 			// Absolute path for resolving entry points and loaders from configuration. 
 			// By default, webpack uses the current directory, but it is recommended to 
 			// manually provide a value.
 			context : path.resolve(__dirname),
-			externals: computeExternals(isRenderer),
 			// Dependencies to exclude from the output bundles.  Instead, the created 
 			// bundle relies on that dependency to be present in the consumer's environment.
 			// This feature is typically most useful to library developers.
@@ -116,15 +78,34 @@ function base(options){
 					__static: `"${path.join(__dirname, "static").replace(/\\/g, "\\\\")}"`,
 					"process.env.NODE_ENV": IS_PRODUCTION ? "\"production\"" : "\"development\""
 				}),
-				new CopyWebpackPlugin({
-					patterns: [ { from: 'static' } ]
-				}),
 				PluginSkeletonOptimization
 			],
-
+			// Determine how the different types of modules within a project will be treated.
+			module: {
+				// An array of rules which are matched to requests when modules 
+				// are created, which can odify how modules are created, apply
+				// loaders to the module, or modify the parser.
+				rules: [
+					{
+						test: /\.node$/,
+						use: "node-loader"
+					},
+					{
+						test: /\.ts$/,
+						exclude: /(node_modules)/,
+						use: [{
+							"loader": "ts-loader",
+							"options": {
+								"transpileOnly": false,
+								"configFile": path.resolve(__dirname, "tsconfig.json")
+							}
+						}]
+					},
+				]
+			},
 			/* TODO (2021/3/1) disable source-map in production mode */
 			devtool: "source-map"
-		});
+		}, options);
 
 		return result;
 	}
