+++ webpack.renderer.js
@@ -77,7 +77,7 @@ function generateIndexFile(nodeModulePath /* string|null */, template /* string 
 
 	if (nodeModulePath) { html = html.replace("</head>", `<script>require('module').globalPaths.push("${nodeModulePath.replace(/\\/g, "/")}")</script></head>`) }
 
-	html = html.replace("</head>", '<script>require("source-map-support/source-map-support.js").install()</script></head>')
+	//html = html.replace("</head>", '<script>require("source-map-support/source-map-support.js").install()</script></head>')
 
 	//if (scripts.length) { html = html.replace("</head>", `${scripts.join("")}</head>`); }
 	//if (css.length)     { html = html.replace("</head>", `${css.join("")}</head>`);     }
@@ -92,7 +92,7 @@ let template = fsExtra.readFileSync(PACKAGE_JSON.electronWebpack.renderer.templa
 ////////////////////////////////////////////////////////////////////////////////
 
 const config = base({
-	target: "electron-renderer",
+	target: "web",
 	// Absolute path for resolving entry points and loaders from configuration. 
 	// By default, webpack uses the current directory, but it is recommended to 
 	// manually provide a value.
@@ -100,12 +100,11 @@ const config = base({
 	// The point or points to start the bundling process.
 	// (if an array is passed, all items will be processed)
 	entry : {
-		renderer: "./src/renderer/index.ts"
+		renderer: "./src/renderer/index.ts",
 	},
 	output: {
 		filename: '[name].js',
 		chunkFilename: '[name].bundle.js',
-		libraryTarget: 'commonjs2',
 		path: path.resolve(__dirname, "dist/renderer")
 	},
 	// Choose a style of source mapping to enhance the debugging process.
@@ -113,7 +112,7 @@ const config = base({
 	// Options for changing how modules are resolved.
 	resolve: { 
 		// aliases to import more easily from common folders
-		alias: { "@" : path.resolve(__dirname, "src/main") },
+		alias: { "@" : path.resolve(__dirname, "src/renderer") },
 		// attempt to resolve file extensions in this order
 		// (allows leaving off the extension when importing)
 		extensions: [".js", ".ts", ".tsx", ".json", ".css", ".node"],
@@ -159,7 +158,7 @@ const config = base({
 							[ require("@babel/preset-env").default, 
 								{
 									modules: false,
-									targets: { electron: "9.1.1" } // TODO: update node version here
+									targets: { electron: "12.0.0" }
 								}
 							],
 							require("@babel/preset-typescript").default,
@@ -171,10 +170,6 @@ const config = base({
 						]
 					}
 				}
-			},
-			{
-				test: /\.node$/,
-				use: "node-loader"
 			},
 			{
 				test: /\.css$/i,
@@ -194,17 +189,6 @@ const config = base({
 					}
 				}
 			},
-			{
-				test: /\.ts$/,
-        		exclude: /(node_modules)/,
-				use: [{
-					"loader": "ts-loader",
-					"options": {
-						"transpileOnly": false,
-						"configFile": path.resolve(__dirname, "tsconfig.json")
-					}
-				}]
-			},
 			// using SolidJS requires custom babel plugin
 			{
 				test: /\.tsx$/,
@@ -216,7 +200,7 @@ const config = base({
 						presets: [
 							[ '@babel/preset-env', { 
 								// TODO: update electron version here
-								"targets": { "electron": "9.0.2", }
+								"targets": { "electron": "12.0.0", }
 							} ],
 							'solid',
 							'@babel/preset-typescript'
